feat(refactor): refactoring user library and user routes
