+++ src/features/users/currently-reading/CurrentlyReading.type.ts
@@ -1,0 +1,7 @@
+export interface ICurrentlyReading {
+  manga_id: string;
+  user_id: string;
+  current_chapter: string;
+  finished_chapters: string[];
+  createdAt: Date;
+}
