+++ src/features/users/user-mangas/UserManga.type.ts
@@ -1,0 +1,12 @@
+export interface IUserManga {
+  manga_id: string;
+  image: string;
+  referer: string;
+  title: string;
+  views: number;
+  todayViews: {
+    date: Date;
+    count: number;
+  };
+  createdAt: Date;
+}
