+++ src/features/users/user/User.type.ts
@@ -1,0 +1,9 @@
+export interface IUser {
+  username: string;
+  email: string;
+  password: string;
+  profilePicture: string;
+  favorites: string[];
+  createdAt: Date;
+  updatedAt: Date;
+}
