--- src/routes/index.ts
@@ -1,2 +1,0 @@
-export { default as UserRoutes } from "./user";
-export { default as MangaRoutes } from "./manga";
