+++ src/utils/setupCors.ts
@@ -1,0 +1,10 @@
+import fastifyCors from "@fastify/cors";
+
+const setupCors = async (fastify: import("fastify").FastifyInstance) => {
+  await fastify.register(fastifyCors, {
+    origin: "*",
+    methods: "GET",
+  });
+};
+
+export default setupCors;
